requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
langchain-google-genai
langgraph
langchain
httpx
sqlmodel
langsmith